0. Migrate from capabilities to just components (much simpler) 1. Gradle upgrade 2. Fix cooldown in summoner spells (int -> double)

+++ app/src/main/java/com/tsarsprocket/reportmid/riotapi/ddragon/SpellDataDto.kt
@@ -3,7 +3,7 @@ package com.tsarsprocket.reportmid.riotapi.ddragon
 import com.google.gson.annotations.SerializedName
 
 data class SpellDataDto(
-    @SerializedName("cooldown") val cooldown: List<Int>,
+    @SerializedName("cooldown") val cooldown: List<Double>,
     @SerializedName("cooldownBurn") val cooldownBurn: String,
     @SerializedName("cost") val cost: List<Int>,
     @SerializedName("costBurn") val costBurn: String,
